♻️ 最低限読めるくらいには分割
