+++ main/F446-INUWASHI/src/process/UI.h
@@ -1,0 +1,12 @@
+#ifndef _UI_H_
+#define _UI_H_
+
+#include "../device/device.h"
+#include "./location.h"
+
+extern Location location;
+
+void UI_settingWrite();
+void UI_runningWrite();
+
+#endif
