+++ main/F446-INUWASHI/src/process/process.h
@@ -1,5 +1,5 @@
-#ifndef _NEKO_H_
-#define _NEKO_H_
+#ifndef _PROCESS_H_
+#define _PROCESS_H_
 
 #include "./location.h"
 extern Location location;
@@ -7,5 +7,7 @@ extern Field Map[200];
 
 #include "./victim.h"
 extern VICTIM victim;
+
+#include "./UI.h"
 
 #endif
